+++ src/state/sagas/test/moveValidationSaga.test.ts
@@ -1,5 +1,5 @@
 import { expectSaga } from 'redux-saga-test-plan';
-import playerMovedSaga from '../MoveValidationSaga';
+import playerMovedSaga from '../moveValidationSaga';
 import { getCurrentPlayer } from '../../selectors/AppStateSelectors';
 import { Player } from '../../AppState';
 import { CHANGE_PLAYER, CHECK_GAME_FINISHED, PLAYER_MOVED } from '../../game/gameAction';
@@ -8,9 +8,8 @@ import { REGISTER_MOVE } from '../../moves/moveAction';
 import { CALCULATE_BOARD_VALUE, SET_TILE_VALUE } from '../../board/boardActions';
 import { CALCULATE_ALLOWED_BOARDS } from '../../activeBoards/activeBoardActions';
 
-// TODO: rename to lowercase (test and actual file)
 // TODO: boardPoint / bigBoardPoint / boardPosition: einheitlicher Name. (Code Refactoring)
-describe( 'MoveValidationSaga', () => {
+describe( 'moveValidationSaga', () => {
     it( 'should dispatch at least the following actions, order not tested', () => {
         return expectSaga( playerMovedSaga )
             .provide( [
